Increase accuracy of break-even calculator
Calculations with an initial upvote count of 250 will be accurate to 1 decimal place, but if it is anything bigger then calculations will be rounded to the nearest whole number.

Processing anything above an initial upvote count of 250 to 1 decimal place starts getting slow. This could be fixed in the future by developing a function to get the break-even point algorithmically, instead of sweeping with a loop.